Define AUTH_ME endpoint constant and use in token refresh
- Add AUTH_ME constant to Constants.h pointing to /api/v1/auth/me
- Update refreshAuthToken() to use the constant instead of hardcoded string
- Verified backend supports GET /api/v1/auth/me endpoint with auth middleware

The backend endpoint returns current user data when token is valid,
and returns 401 when token is expired/invalid, making it perfect for
token validation in the refresh flow.

+++ plugin/src/security/RateLimiter.cpp
@@ -96,7 +96,7 @@ void TokenBucketLimiter::cleanup() {
     }
 
     // Stop if we're under max size
-    if (static_cast<size_t>(buckets_.size()) < config_.maxTrackedIdentifiers)
+    if (static_cast<size_t>(buckets_.size()) < static_cast<size_t>(config_.maxTrackedIdentifiers))
       break;
   }
 
@@ -227,7 +227,7 @@ void SlidingWindowLimiter::cleanup() {
       keysToRemove.add(it.getKey());
     }
 
-    if (static_cast<size_t>(windows_.size()) < config_.maxTrackedIdentifiers)
+    if (static_cast<size_t>(windows_.size()) < static_cast<size_t>(config_.maxTrackedIdentifiers))
       break;
   }
